Refactored DCCMain and DCCService classes into one, refactored RailCom
